Use @roo-code/build in src/esbuild.mjs
