+++ apps/vscode-nightly/package.json
@@ -4,11 +4,11 @@
 	"private": true,
 	"packageManager": "pnpm@10.8.1",
 	"scripts": {
-		"build": "rimraf build && pnpm --filter @roo-code/build build && node esbuild.mjs --production && pnpm --filter @roo-code/vscode-webview build --mode nightly",
+		"build": "pnpm clean && pnpm --filter @roo-code/build build && node esbuild.mjs --production && pnpm --filter @roo-code/vscode-webview build --mode nightly",
 		"vsix": "pnpm build && cd build && mkdirp ../../../bin && npx vsce package --no-dependencies --out ../../../bin",
-		"clean": "rimraf build"
+		"clean": "rimraf build .turbo"
 	},
-	"dependencies": {
+	"devDependencies": {
 		"@roo-code/build": "workspace:^"
 	}
 }
