+++ src/esbuild.mjs
@@ -1,0 +1,163 @@
+import * as esbuild from "esbuild"
+import * as fs from "fs"
+import * as path from "path"
+import { fileURLToPath } from "url"
+
+import { copyPaths, copyWasms, copyLocales } from "@roo-code/build"
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+
+function setupLocaleWatcher() {
+	if (!watch) {
+		return
+	}
+
+	const localesDir = path.join(__dirname, "i18n", "locales")
+
+	if (!fs.existsSync(localesDir)) {
+		console.warn(`Cannot set up watcher: Source locales directory does not exist: ${localesDir}`)
+		return
+	}
+
+	console.log(`Setting up watcher for locale files in ${localesDir}`)
+
+	let debounceTimer = null
+
+	const debouncedCopy = () => {
+		if (debounceTimer) {
+			clearTimeout(debounceTimer)
+		}
+
+		// Wait 300ms after last change before copying.
+		debounceTimer = setTimeout(() => {
+			console.log("Locale files changed, copying...")
+			copyLocaleFiles()
+		}, 300)
+	}
+
+	try {
+		fs.watch(localesDir, { recursive: true }, (eventType, filename) => {
+			if (filename && filename.endsWith(".json")) {
+				console.log(`Locale file ${filename} changed, triggering copy...`)
+				debouncedCopy()
+			}
+		})
+		console.log("Watcher for locale files is set up")
+	} catch (error) {
+		console.error(`Error setting up watcher for ${localesDir}:`, error.message)
+	}
+}
+
+async function main() {
+	const production = process.argv.includes("--production")
+	const watch = process.argv.includes("--watch")
+	const minify = production
+	const sourcemap = !production
+
+	/**
+	 * @type {import('esbuild').BuildOptions}
+	 */
+	const buildOptions = {
+		bundle: true,
+		minify,
+		sourcemap,
+		logLevel: "silent",
+		format: "cjs",
+		sourcesContent: false,
+		platform: "node",
+	}
+
+	const srcDir = __dirname
+	const buildDir = __dirname
+	const distDir = path.join(buildDir, "dist")
+
+	/**
+	 * @type {import('esbuild').Plugin[]}
+	 */
+	const plugins = [
+		{
+			name: "copy-files",
+			setup(build) {
+				build.onEnd(() => {
+					copyPaths(
+						[
+							["../README.md", "README.md"],
+							["../CHANGELOG.md", "CHANGELOG.md"],
+							["../LICENSE", "LICENSE"],
+							["node_modules/vscode-material-icons/generated", "assets/vscode-material-icons"],
+							["../webview-ui/audio", "webview-ui/audio"],
+						],
+						srcDir,
+						buildDir,
+					)
+				})
+			},
+		},
+		{
+			name: "copy-wasms",
+			setup(build) {
+				build.onEnd(() => copyWasms(srcDir, distDir))
+			},
+		},
+		{
+			name: "copy-locales",
+			setup(build) {
+				build.onEnd(() => copyLocales(srcDir, distDir))
+			},
+		},
+		{
+			name: "esbuild-problem-matcher",
+			setup(build) {
+				build.onStart(() => console.log("[esbuild-problem-matcher#onStart]"))
+				build.onEnd((result) => {
+					result.errors.forEach(({ text, location }) => {
+						console.error(`✘ [ERROR] ${text}`)
+						console.error(`    ${location.file}:${location.line}:${location.column}:`)
+					})
+
+					console.log("[esbuild-problem-matcher#onEnd]")
+				})
+			},
+		},
+	]
+
+	/**
+	 * @type {import('esbuild').BuildOptions}
+	 */
+	const extensionConfig = {
+		...buildOptions,
+		plugins,
+		entryPoints: ["extension.ts"],
+		outfile: "dist/extension.js",
+		external: ["vscode"],
+	}
+
+	/**
+	 * @type {import('esbuild').BuildOptions}
+	 */
+	const workerConfig = {
+		...buildOptions,
+		entryPoints: ["workers/countTokens.ts"],
+		outdir: "dist/workers",
+	}
+
+	const [extensionCtx, workerCtx] = await Promise.all([
+		esbuild.context(extensionConfig),
+		esbuild.context(workerConfig),
+	])
+
+	if (watch) {
+		await Promise.all([extensionCtx.watch(), workerCtx.watch()])
+		copyLocaleFiles()
+		setupLocaleWatcher()
+	} else {
+		await Promise.all([extensionCtx.rebuild(), workerCtx.rebuild()])
+		await Promise.all([extensionCtx.dispose(), workerCtx.dispose()])
+	}
+}
+
+main().catch((e) => {
+	console.error(e)
+	process.exit(1)
+})
