+++ src/package.json
@@ -320,19 +320,20 @@
 	"scripts": {
 		"lint": "eslint **/*.ts",
 		"check-types": "tsc --noEmit",
-		"pretest": "node esbuild.js",
+		"pretest": "node esbuild.mjs",
 		"test": "jest -w=40% && vitest run",
 		"format": "prettier --write .",
-		"build": "pnpm lint && pnpm check-types && rimraf dist && node esbuild.js --production && rimraf webview-ui && pnpm --filter @roo-code/vscode-webview build",
-		"build:development": "pnpm lint && pnpm check-types && rimraf dist && node esbuild.js && rimraf webview-ui && pnpm --filter @roo-code/vscode-webview build",
+		"build": "pnpm lint && pnpm check-types && pnpm clean && pnpm --filter @roo-code/build build && node esbuild.mjs --production && pnpm --filter @roo-code/vscode-webview build",
+		"build:development": "pnpm lint && pnpm check-types && pnpm clean && pnpm --filter @roo-code/build build&& node esbuild.mjs && pnpm --filter @roo-code/vscode-webview build",
 		"publish:marketplace": "vsce publish --no-dependencies && ovsx publish --no-dependencies",
 		"publish": "pnpm vsix && changeset publish && pnpm install --package-lock-only",
 		"version-packages": "changeset version && pnpm install --package-lock-only",
 		"vscode:prepublish": "pnpm build",
 		"vsix": "mkdirp ../bin && npx vsce package --no-dependencies --out ../bin",
-		"watch:esbuild": "node esbuild.js --watch",
+		"watch:esbuild": "node esbuild.mjs --watch",
 		"watch:tsc": "tsc --noEmit --watch --project tsconfig.json",
-		"generate-types": "tsx scripts/generate-types.mts"
+		"generate-types": "tsx scripts/generate-types.mts",
+		"clean": "rimraf README.md CHANGELOG.md LICENSE dist webview-ui out .turbo"
 	},
 	"dependencies": {
 		"@anthropic-ai/bedrock-sdk": "^0.10.2",
@@ -398,6 +399,7 @@
 		"@changesets/cli": "^2.27.10",
 		"@changesets/types": "^6.0.0",
 		"@jest/globals": "^29.7.0",
+		"@roo-code/build": "workspace:^",
 		"@types/clone-deep": "^4.0.4",
 		"@types/debug": "^4.1.12",
 		"@types/diff": "^5.2.1",
